mocktikv: extract test utils (#24066)
